Develop

+++ API/Controllers/UsersController.cs
@@ -1,6 +1,7 @@
 using Microsoft.AspNetCore.Mvc;
 using CryptoTrade.Business;
 using CryptoTrade.Models;
+using Microsoft.AspNetCore.Authorization;
 
 namespace CryptoTrade.API.Controllers;
 
@@ -9,12 +10,15 @@ namespace CryptoTrade.API.Controllers;
 public class UsersController : ControllerBase
 {
     private readonly IUserService _userService;
+    private readonly IAuthService _authService;
 
-    public UsersController(IUserService userService)
+    public UsersController(IUserService userService, IAuthService authService)
     {
         _userService = userService;
+        _authService = authService;
     }
 
+    [Authorize(Roles = Roles.Admin)]
     [HttpGet]
     public ActionResult<IEnumerable<User>> GetAllUsers()
     {
@@ -29,9 +33,13 @@ public class UsersController : ControllerBase
         }
     }
 
+    [Authorize(Roles = Roles.Admin + "," +  Roles.User)]
     [HttpGet("{userId}")]
     public IActionResult GetUser(int userId)
     {
+        if (!_authService.HasAccessToResource(Convert.ToInt32(userId), null, HttpContext.User)) 
+            {return Forbid(); }
+
         try
         {
             var user = _userService.GetUserById(userId);
@@ -47,9 +55,13 @@ public class UsersController : ControllerBase
         }
     }
 
+    [Authorize(Roles = Roles.Admin + "," +  Roles.User)]
     [HttpGet("by-email")]
     public IActionResult GetUserByEmail(string userEmail)
     {
+        if (!_authService.HasAccessToResource(null, userEmail, HttpContext.User)) 
+            {return Forbid(); }
+
         try
         {
             var user = _userService.GetUserByEmail(userEmail);
@@ -65,12 +77,17 @@ public class UsersController : ControllerBase
         }
     }
 
+    [Authorize(Roles = Roles.Admin + "," +  Roles.User)]
     [HttpPut("{userId}")]
     public IActionResult UpdateUser(int userId, UserUpdateDTO userUpdateDTO)
     {
         if (!ModelState.IsValid)  {return BadRequest(ModelState); } 
 
-        try {
+        if (!_authService.HasAccessToResource(Convert.ToInt32(userId), null, HttpContext.User)) 
+            {return Forbid(); }
+
+        try 
+        {
             _userService.UpdateUser(userId, userUpdateDTO);
             return Ok("Usuario actualizado correctamente.");
         }     
@@ -84,9 +101,13 @@ public class UsersController : ControllerBase
         }
     }
 
+    [Authorize(Roles = Roles.Admin + "," +  Roles.User)]
     [HttpDelete("{userId}")]
     public IActionResult DeleteUser(int userId)
     {
+        if (!_authService.HasAccessToResource(Convert.ToInt32(userId), null, HttpContext.User)) 
+            {return Forbid(); }
+
         try
         {
             _userService.DeleteUser(userId);
